Move template engine related code from lib/utils to lib/render
