+++ lib/utils.js
@@ -1,14 +1,12 @@
 // Core dependencies
 const crypto = require('crypto')
 const fs = require('fs')
+const path = require('path')
 
 // NPM dependencies
 const { get: getKeypath } = require('lodash')
-const { marked } = require('marked')
-const path = require('path')
 const portScanner = require('portscanner')
 const inquirer = require('inquirer')
-const matter = require('gray-matter')
 
 // Are we running on Glitch.com?
 const onGlitch = function () {
@@ -31,20 +29,6 @@ exports.getNodeEnv = function () {
   const env = (process.env.NODE_ENV || glitchEnv || 'development').toLowerCase()
   return env
 }
-
-// Tweak the Markdown renderer
-const defaultMarkedRenderer = marked.defaults.renderer || new marked.Renderer()
-
-marked.use({
-  renderer: {
-    code (code, infostring, escaped) {
-      let rawHtml = defaultMarkedRenderer.code(code, infostring, escaped)
-      // Add a tabindex to the <pre> element, to allow keyboard focus / scrolling
-      rawHtml = rawHtml.replace('<pre>', '<pre tabindex="0">')
-      return rawHtml
-    }
-  }
-})
 
 // Require core filters and then add the methods to Nunjucks environment
 exports.addNunjucksFilters = function (env) {
@@ -150,32 +134,6 @@ exports.forceHttps = function (req, res, next) {
   // Mark proxy as secure (allows secure cookies)
   req.connection.proxySecure = true
   next()
-}
-
-exports.isMdView = function (mdDir, name) {
-  const reqPath = path.join(mdDir, name + '.md')
-  return fs.existsSync(reqPath)
-}
-
-exports.markdownEngine = (nunjucksEnv) => {
-  return (filePath, options, callback) => {
-    fs.readFile(filePath, (err, fileContents) => {
-      if (err) return callback(err)
-      const parsedFile = matter(fileContents)
-      if (parsedFile.data.heading === undefined) {
-        throw new Error(`${filePath} does not have a heading in its frontmatter`)
-      }
-      const layout = parsedFile.data.layout || 'documentation_template.html'
-      marked(parsedFile.content, (err, html) => {
-        if (err) return callback(err)
-        nunjucksEnv.render(
-          layout,
-          { ...options, ...parsedFile.data, document: html },
-          callback
-        )
-      })
-    })
-  }
 }
 
 const sleep = (ms) => new Promise((resolve) => setTimeout(resolve, ms))
